Renamed index page and updated readme
