+++ cat-shark-carousel/src/index.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import Index from './pages/Index';
+import Home from './pages/Home';
 
 ReactDOM.render(
   <React.StrictMode>
-    <Index />
+    <Home />
   </React.StrictMode>,
   document.getElementById('root')
 );
