+++ cat-shark-carousel/src/pages/Home.tsx
